Handle null and empty values across inputs
Improves data safety and validation logic by ensuring proper handling of null or empty values in inputs and calculations. Enhances app stability and prevents potential errors caused by unexpected input conditions. strlen(null) & trim(null) are deprecated.

+++ src/Api/SeriesUpdates.php
@@ -260,9 +260,9 @@ class SeriesUpdates extends AbstractController
             $data = $this->getDates($progress);
         }
 
-        $lastUpdate = $this->dateService->newDateFromTimestamp(($data['end date'] / 1000) ?? 0, "UTC")->format("Y-m-d H:i:s");
+        $lastUpdate = $this->dateService->newDateFromTimestamp((int)($data['end date'] / 1000) ?? 0, "UTC")->format("Y-m-d H:i:s");
         $lastUpdateString = $this->dateService->formatDateRelativeLong($lastUpdate, "Europe/Paris", $request->getLocale());
-        $lastDuration = ($data['end date'] - $data['start date']) / 1000;
+        $lastDuration = (int)($data['end date'] - $data['start date']) / 1000;
         $lastDurationString = $this->dateService->getDurationString($lastDuration, $units);
 
         return new JsonResponse([
